pyproject.toml
[tool.poetry]
name = "gohobiticket2"
version = "0.1.0"
description = ""
authors = ["lynxlevin <72173230+lynxlevin@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.2.11"
djangorestframework = "^3.13.1"
Markdown = "^3.3.7"
django-filter = "^21.1"
psycopg2 = "^2.9.3"
urllib3 = "^1.26.9"
django-environ = "^0.9.0"
mypy = "^0.971"
requests = "^2.28.1"
isort = "^5.10.1"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
factory-boy = "^3.2.1"
types-requests = "^2.31.0"

[tool.black]
line-length = 160
target-version = ['py36']
exclude = '''
/(
    \.git
    | \.mypy_cache
    | \.venv
    | \.vscode
)/
'''

[tool.isort]
include_trailing_comma = true
line_length = 160
multi_line_output = 3

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
